Swing the infamous "phk viking axe" and deorbit directors from the core code.
In the future directors belong in VMODs, which takes the need to
hack VCC code out of the learning-curve.

Certain directors have already been moved to VMODs, others will
need a bit of brain-surgery before they go there.

This commit is to get them out of the way for V4