Turn obj.hits and obj.last_use (notice new '_', see below) into VCL only variables, and add a commented example in default.vcl.
The reason for this is to avoid VM write-backs to cached objects by default.

Previously we did this with the obj_readonly parameter, but this puts
it right where the user can see it.

Notice that the two variables are not locked, so they cannot be
trusted to be precise.

obj.lastuse changes to obj.last_use because its type changes from
DURATION to TIME, (and to improve readability in general)

Fix & Polish relevant test-cases