Move query input to be an option instead of cmd line args
Having the query input to the utils be cmd line args leads users to
believe that the args are interpreted as tokens themselves. The parser
doesn't see each argument as a token, but instead makes one string
from all the arguments. This would cause quoting issues.

By having the query expression be an option instead, it becomes
clearer that the input is one string only, and any quoting when
constructing an expression from some other input is left to the user.