Totally overhaul the EXP code.
In general new/changed/killed objects are posted in the exp_thread
mailbox, and suitably handled by the exp_thread at its convenience.

The binheap is only manipulated by exp_thread, and it doesn't
need to hold any locks to do so.

The exp_mtx only protects the inbox to the exp_thread.

oc->flags and the lru-list are protected by the lru->mtx.

EXP_NukeLRU() is temporarily disabled.

When LRU-Nuking to make space, we mark the object as dying, so HSH
will ignore it and steal the storage from it right away.  We post the
rest of the object to exp_thread to clean up later.

The LRU change makes the 3-4 LRU test-cases unstable, (c00044, c00045,
r01140...) because they were used to the full object being reaped,
rather than only the storage segments.  I'm to tired to fix/tune those
cases tonight.

But apart from that, it seems to run the test-cases.