VSLQ_Dispatch now returns sooner.
Change VSLQ_Dispatch to always return after processing one record or
one batch of records. It will return 1 if it should be called again
without sleeping.

This fixes the situation where a calling program is lagging behind,
and VSLQ_Dispatch never returns because there is a constant stream of
new log records.