Add batch records as a cursor option to the VSM cursor
If the option isn't set, batch records are never returned from the cursor.

If the option is set, batch records are reported, and the next call to
the cursor will return the first record after the batch. Calling
application will then need to know how to use VSL_NEXT pointer macros
to extract the records in the batch.