Renovate the parameter/cli code.
Certain of our parameter names are so long that they have shoved the
"long" param.show output all the way to the right.  Break the line
after the param name and use most of the width again.

Also add support for two column tables using TAB chars, for descriptions
of params like "debug", "feature" etc.  The width autosizes, and lines
wrap as expected in the second column.  Eliminate the 'pad' field in
the tbl includes.

Move all the "tweak" functions to a separate source file.

Clean up white-space in param descriptions.