The corresponding API changes
This makes the logging API work again after the changes to the Begin
and Link log records.

Even though restarts are linked below the request that initiated them,
the API will report them at the same level as the initiating
request. This is to help makes queries depending on levels
predictable.