Rename default.vcl to builtin.vcl.
The default.vcl name was thought up in 2007 and have been a point of confusion
since.

New users take the old default VCL, uncomment what they think makes sense, and
always end with return(lookup) because that seems to be needed. But it isn't,
and the defaults in the built-in VCL are much better at handling all things
they haven't learned about yet.