Remove extra meta-header.
We never know when to update this header. All the important parts of it is
available in the git log. Since we don't really do versioning of documentation,
it was really just taking up space.

I've tried my best to keep the author names from the headers in the history
section.