Add manual Provides for libvarnishapi.so.1.
find-provides did not pick up on libvarnishapi.so.1 for varnish-libs
automatically, and find-requires for the main package did. This meant
that package pair was uninstallable.

I don't quite know why this is happening, so I'm working around it
with a hardcoded Provides for now. This is on the critical path for
4.0 release.