Be more RFC2616 compliant, by keeping Date header as given from the backend. Set it on the object if missing.
Previously we would add an explicit Date header on each response with
the time and date of the response, overwriting any Date header on the
object that came from the backend.

Now we will add a Date header if the backend didn't supply one just
before vcl_backend_response() is run, and not add any Date headers for
the replies.

This change means Date headers sent by Varnish is approximately the
time and date this object was generated by the backend, and not the
current date and time.