Update the documentation for obj.hits to reflect code reality.
In 4.x we changed obj.hits to be read-only, and to count all variants
under the same hash-key (= same objhdr).

Since we cannot trivially tell if a ban hits all Vary: values or
just some of them, there is no sane heuristic to fiddle obj.hits
on bans that would Do The Right Thing.

Fixes #1492