Add option parser, error handling and strictness.
This is one of those tiny jobs you start on (add --strict),
and then it just goes downhill from there.

Main changes:
* Use optparse to parse command line arguments.
* With --strict it aborts if a parse or format error is seen.
* A basic test case for the parser has been added.

No changes were made to the parser or output format.

Adding the test case meant reordering the execution flow.

Since Varnish code standard is using tabs, I've kept this to my
best ability. Pylint is not happy about this and give the code
a score of -4 out of 10 points. :-)