Give the CLI command vcl.show a -v flag which outputs all the source files involved in a given VCL.
When using -v, the individual source files are output like this:

	// VCL.SHOW %d %d %s\n%s\n

First field is the source file index [0...].
Second field is the number of bytes in the source file.
Third field is the name of the source file.
Fourth field is the source file content.

Inspired by a patch from daghf