Rename shm_reclen to vsl_reclen for consistency.
Leave shm_reclen as a parameter alias for now.

Parameter vsl_buffer must be 12 bytes larger than vsl_reclen in order
to avoid a panic when we try to put 12 pounds of VSL into a 5 pound
vsl_buffer sack.   Tweak the opposite parameter Minimum or Maximum value
when we set one of of these parameters.

Fixes #1547