Fix sensitivity for header token matching. Rename GetHdrData -> GetHdrToken
Section 4.2 (Messages Headers) of RFC2616 defines field (header) name as case
insensitive, but the field (header) value/content may be case-sensitive.

http_GetHdrToken looks up a token in a header value and the rfc does not say
explicitly if tokens are to be compared with or without respect to case.

But all examples and specific statements regarding tokens follow the rule
that unquoted tokens are to be matched case-insensitively and quoted tokens
case-sensitively.

Fixes #1582