Do not log garbage requests.
Requests that end up in the hard "400 Bad Request" handling used to be
logged with incomplete data. (no method, maybe no path, possibly no proto,
and no response status)

Port scans or anything sending a byte and a linefeed would be logged.

Since this is used for logging access to a web site, it makes more sense
to skip these garbage requests.

Fixes:		#1584