Firmly split "director" from "backend".
A director is noe an abstract entity which can fetch something for us,
however it decides to do that and using whatever protocol (or none).

A backend is a network endpoint designated by a precise triplet of
	VCL backend name
	IPv4 address (incl port)
	IPv6 address (incl port)

Backends are _also_ "simple" directors, but have an extra layer of
complexity in order to enable connection management, including
connection reuse across vcl.use commands.