Add a VFIL_allocate function
This function makes sure that the filesystem can accomodate the file
of the given size.

Checks the filesystem free space up against the allocated blocks as a
sanity check.

Uses fallocate if available to preallocate the needed blocks.

If fallocate isn't available, it will write zero bytes if insist is true.