varnishlog -k argument
Implement -k argument in vut for use with varnishlog. This takes as an
argument the number of log transactions to be processed before
exiting. Useful to stop varnishlog after a given set of log data has
been produced.

Drop -s from the 'not implemented yet' list as this option has very
limited usefullness.