Convert (and retire) the sandbox code to the new jail framework.
The retires the -u and -g arguments, and the user/group/group_cc
parameters, so that the jail-setup cannot be manipulated from CLI at all.

$user and group_cc are now sub-arguments to -junix, for instance:

	-junix,user=varnish1,ccgroup=ccowner

There is no group= subargument, the group information is taken
from the user id in question.

If no -j argument is specified, attempt:

	-junix,user=varnish

but fail silently if not possible.

The "vident" now contains information about jail-config.

Sandbox_solaris is not converted yet, (I'm hoping that Nils will
help with that.