Put a protocol prefix on enum sess_step.
In the future we will need to tell what protocol controls a give
session, for instance when we come back from the waiter or waiting
list.   Rather than spend memory on a dedicated field for this,
protocols will share the enum sess_step, and ranges will be used
to dispatch to protocols.