Run background processes in varnishtest
Available operations on processes are:
- start
- wait
- kill (send a signal)
- stop (kill with TERM)
- write (to its stdin)
- close (its stdin)

A process name starts with a 'p' and accepts a shell pipeline as the
command line. Environment variables can be set directly in the command
line.

A process will create a macro with its PID, and its stdout and stderr
are redirected to files in a dedicated directory.

Also add a varnish "name" macro in varnishtest.