Reformat JSON output for 80 character width.
This output was very wide and hard to get any sense out of without
actually reading it closely.

The things you are interested in, typically counter name, description
and value, are now on separate lines.

In addition the output should now fit nicely on any VT100 terminals in
80 character mode, I'm sure there are some still in use somewhere.

JSON document layout (key names, hierarchy) is the same as before.