Move the built HTML documentation to doc/html/.
I think it makes sense to ship a built HTML copy in
the tarball, as well as the RST sources.

The old cleanup code was a tad eager and made phk miss
his svg. Hopefully this way is a bit more future proof.

automake is our friend.

We might have to update some packaging files later on.
(and perhaps the git push hook.)