Increase libvarnishapi libtool revision number.
This number was last changed between Varnish 1.0 and Varnish 2.0.

As far as I can read from the commits, we haven't actually removed
any interfaces in any released Varnish versions, so a revision bump
should be all that is needed.

Another possibility is to stop using -version-info and just roll
our own scheme with -release, but I don't want to go down that route
just before 4.1.0.