Allow invalid headers in 304 backend responses.
Allow the backend server to send headers lacking ":"/colon in
them when responding to a conditional request yielding a 304 response.

In master/4.1 such responses are aborted as invalid. The backend is
clearly not feeling well.

Since we've accepted it nicely for 200 responses so far in Varnish 4.0,
continue that trend also for 304s.

Fixes: #1598