Log proxy related messages on the session, not on the request.
The proxy log records were attempted to be logged on the request' log
buffer, which at that point in time has not yet been set up and does
not posess a VXID. This caused VXID==0 log records to be inserted in
the beginning of the log for the first request on the session when
using proxy protocol, and subsequently the VSL to fail picking them
out as valid request log transactions.

Fix by logging the PROXY handling related messages on the session, in
which they belong.

Fixes: #1804