Implement three different types of attributes and generate get/set code from table
Fixed size attributes has it's space always reserved.

Variable size attributes change in size, and their combined total
length needs to be given to allocobj.

Auxiliary attributes do not need to be predeclared.

The code for setting and getting the attributes is generated from a
table that lists the different attributes according to their type.

Add documentation about the behaviour of the get/set functions.