So one of those strange cornercases in HTTP/1
If we send the backend a HTTP/1.0 request, and it doesn't have a
Content-Length, it cannot use Chunked and must fall back to EOF.

However, the protocol field in the response tells us what version
backend *could* have used, not what it *does* use.

So we can get a response with HTTP/1.1 and EOF, following HTTP/1.0
semantics - because we asked for it.

Most sensible backends avoid this, either by buffering and creation
of a C-L or, smartly, returning "HTTP/1.0", even though that
is strictly speaking against the apocrphal texts.

Anyway, now we cope...

Fixes: 	#1810