Add a final backstop, so we absolutely 100% certainly do not try to delete a objhead while it still has a waiting list, by forcing the last ref holder to rush the WL.
Since the hasher owns the refcounts on objhead, we cannot just
mingle req and objcore refcounts.

Fortunately we don't need to add another refcounter, because all
we really care about is the wl being empty when we drop the last
ref.

The wl/hsh_rush() mechanism will work differently with different
thread-scheduling schenarios, and I cannot definitively rule out
that we can drop the last ref on an oh, while there are still req's
on the waiting list.

Given that, and the existence proof in ticket #1823's race, this
might have been the indicated memory-trampler.