WS: Fix off-by-one in WS_Reset
WS_Assert allows the workspace to be full in its check that `ws->f <= ws->e`.
WS_Reset does not honor this; if a snapshot was taken of a full workspace, and
that snapshot is reverted with WS_Reset, we panic.

This issue was identified by Jozef Hatala <jozef@fastly.com>