Split BAN_RefBan into BAN_FindBan and BAN_RefBan
BAN_FindBan looks up a ban based on a ban timestamp and returns a
pointer to the ban. Returns NULL if no ban with that timestamp is
present.

BAN_RefBan takes an objcore and a ban. It grabs a reference to the ban
and associates the objcore with that ban.