Only set new lurker destination ban if we already had one.
Without this we could start running ban_lurker_test_ban with an empty
list of bans to test, causing a lot of unnecessary ban and oh mutex
contention on every ban lurker invocation until the REQ ban triggering
it is gone.

Also don't call ban_lurker_test_ban when our destination ban is the
same as the one being tested. This would be wasted work.

Adjust the c00049.vtc test case for the bans now not being
unnecessarily tested.