Remove assert in shared memory
Even though the shared memory log should always have only one writer,
the varnish process, this is not enforced by the kernel through access
rights. For this reason, the line

     assert(b->vsc->vcls > 0);

could crash Varnish if a process decided to write to the shared memory
log. Now this bad behavior will just affect other readers of the
shared memory log, and not the varnish process itself.