Proper handling of duplicate headers on IMS headers merge
With this patch we properly handle duplicated headers when using http
conditionals (dubbed backend IMS in Varnish). If the backend first
sends the following:

foo: a
foo: b

the backend forwards these as is. If then the objects grace period
ends, but the object is not evicted from the cached, varnish will send
a if-none-match header to the backend. The backend can then reply with
a status 304 Not Modified. In this case the backend can send new 'foo'
headers, or not. Varnish then has to react accordingly by either
sending the original set of 'foo' headers, or the new set of 'foo'
headers. The attached test checks both cases of backend behavior.

Ref: #1879