Introduce "VCL labels".
A VCL label is a symbolic name pointing to a "real" VCL program.

VCL labels can be the active VCL and can be repointed to a different
VCL at any time.

Labels are always warm, and can be discarded with vcl.discard if not
in use.

Labeled VCLs are _also_ always warm.

One possible use could be for site develpment to label a "production",
and a "emergency" VCL so operations personel only have to know
these two labels, not worrying about versioning of the VCLs.

Of course I have other interesting evil plans for this too...