Introduce "VCL labels".
A VCL label is a symbolic name pointing to a "real" VCL program.

Labels are created with "vcl.label <labelname> <vclname>"

Labels are removed with vcl.discard.

VCL labels can be the active VCL and can be repointed to a different
VCL while being so.

Labels and the VCLs they point to are always warm.

One possible use could be for site develpment to label a "production",
and a "emergency" VCL so operations personel only have to know
these two labels, not worrying about versioning of the VCLs.

Of course I have evil plans which are more interesting...