Make hit-for-pass cause HSH_MISS instead of HSH_HIT+OC_F_PASS.
it is more consistent to hit vcl_miss{} every time for these objects,
as opposed to vcl_pass{} most of the time but occationally vcl_miss{}.

This also almost entirely takes the TTL of HFP objects out of the
picture, in the sense that any cacheable object fetched as a result
of HFP will be put into the cache and be usable, rather than the
HFP preventing cacheable objects for whatever TTL the HFP might
have.

See c00076 for an example of this.