Don't trust dlopen refcounting.
Some time ago I decided that it was more convenient if it were
actually possible to locate the compiled VCL shared library based
on the loaded VCL name, and removed the part of the subdirectory
name which made it unique.

Bad idea.

Add a comment to make sure I don't get that Idea again.

Fixes #1933

Merge notes:
* Removed label check, as that doesn't exist in 4.1
* Rewrote commit title.

Conflicts:
	bin/varnishd/mgt/mgt_vcl.c