Ensure we can always set a standard response code
The previous attempt to address 1990 was only half-hearted, the fundamental issue
was that we could still fail to set even a 503 response code because we required
workspace to stringify it.

For well known status codes, we now keep constant strings together with the
default reason phrases and emit those.

Add a vtc for the 1990 case.

Fixes #1990