The parameter's name is timeout_linger
There are two timeout names where `timeout` is the prefix:

- timeout_linger
- timeout_idle

If it were up to me, they'd be called the other way around and according
to the documentation:

> `session_linger` has been renamed to `timeout_linger

If it were up to me, it would have been renamed to a self-describing
`session_linger_timeout` (wink wink, nudge nudge).

Typo spotted by scn.