New assert-like macros `require`, `RN` and `RZ`
There are many places in the code base where assertions are used to
verify an operation relying on a dependency where there is no path for
recovery.

For instance, many `malloc` calls are followed by an `AN` check because
there is no recovery path when the child runs out of memory.  So instead
the classic course of action is to let it panic and rely on the manager
for spawning a new child.

However, `assert` can be turned into a no-op, meaning that an allocation
failure would likely be followed by a segmentation fault slightly harder
to diagnose.  Instead of assimilating dependencies to invariant checks
(assert) the new macros ensure a systematic failures.

These macros should be used for checks on foreign code, free of context.
Assuming that the underlying allocator is bug-free, the only reason to
return a NULL pointer is the lack of memory whereas a failure on a lock
operation would more likely be caused by a misuse of a pthread object
and thus better fits assertions.

Once such patterns are identified they can be applied to the whole code
base conveniently with a Coccinelle semantic patch.