Modernize the build system for out-of-tree vmods
New autoconf macros doing as much heavy lifting as allowed by the
autotools suite.  They work independently of the old macros that now
live in `varnish-deprecated.m4`.  They work nice in multi-module build
trees regardless of whether module sources live in a single directory or
several of them.

Tested on GNU/Linux, SmartOS and FreeBSD although not on all variants of
all supported OSs and architectures.  It doesn't work on el5 that was no
longer supported since the Varnish 4.1 series anyway.

It is a little bit opinionated in order to work with multi-modules, but
nothing that couldn't be improved later, and nothing too intrusive: it
makes use of the `-o` option of vmodtool and doesn't have a mechanism
yet for passing options.  For that it requires at least Varnish 4.1.3
and it announces a support @since Varnish 4.1.4, assuming that if
accepted in 5.0 it would also be made available in 4.1.

It should make the life of VMOD maintainers a lot easier without forcing
them to migrate their build systems right away.  The old macros could be
killed with the next major release.  Ideally they should also build the
`std` and `directors` module because dogfooding. <insert promises of a
unicorns and rainbows and happy princesses here>