Add script to run Coverity.
This will allow anyone with the Coverity scanner installed and our
authtoken to upload their own feature branch if they want to.

Main point is to get this out of the Jenkins job definition and into
something that is more developer accessible.

Doing my best to keep it /bin/sh compatible.