Add a "key" field to each backend.
Normally, backends with the same host (as determined by IPv4/IPv6 address pair)
and port will share TCP connections, but if you give them different keys,
they will not. This is useful if your HTTP connections have some kind of state
(e.g. for security) which makes it impossible to reuse them across backends.

This is a (trivial) forward-port of a 4.1 patch we've been running in production.