Use "none" to disable authentication
This is an adaptation of b2c18706f2d4b100 by Federico G. Schwindt.

Previously you needed an empty argument (''), which didn't play very
well along sysv init scripts and was not documented. Now both "none"
and the empty argument work, but an extra warning is echoed if the
empty argument is used.

The documentation only mentions "none" for disabling the secret, since
this is the only way of doing it in the next version.

Fixes #2007.