remove an assertion which we may easily trip over when the clock is stepped
The assertion itself is semantically correct, but using wallclock time for
this kind of check is inappropriate.

The mid-term plan is to re-work time keeping in varnish and re-check where
using a monotonic clock is more appropriate, but for now we should
at least not panic just because wall clock time was stepped back.

Fixes #1874 for now