mv varnish.m4 varnish-legacy.m4
This file will now contain legacy macros, that is to say shims to deal
with older platforms and macros for out-of-tree VMODs introduced with
Varnish 4.0.0-tp1.

These macros work well but their interfaces aren't ideal and they leave
too much autotools work to downstream VMOD maintainers. They have also
suffered a lack of maintenance, like this 3yo comment:

    # Uses internal function for now..

New macros will be introduced, but the current macros are left as-is and
VMOD maintainers won't be forced to upgrade their build system.