RFC OCD
rfc7231#section-4.3.3

Responses to POST requests are only cacheable when they include
explicit freshness information (see Section 4.2.1 of [RFC7234]).

rfc7231#section-4.3.4

Responses to the PUT method are not cacheable.

rfc5789#section-2

A response to this method is only cacheable if it contains explicit
freshness information (such as an Expires header or "Cache-Control:
max-age" directive) as well as the Content-Location header matching
the Request-URI, indicating that the PATCH response body is a resource
representation.  A cached PATCH response can only be used to respond to
subsequent GET and HEAD requests; it MUST NOT be used to respond to other
methods (in particular, PATCH).