Fix test for duplicate links in vsl_dispatch
This test wasn't working as it was supposed to, and would cause the
forward links graph building of log transactions to fail. The fail was
soft, as the reverse mapping still succeeded creating the right
mapping.

Depending on when the local thread log buffer was flushed, the problem
would show as spurious "duplicate link" log records to be inserted.

Fixes: #1830