Add beresp.storage
This supersedes beresp.storage_hint but is kept around for backward
compatibility until the next major release.

Setting beresp.storage_hint to a valid storage will set beresp.storage
as well. If the storage is invalid, beresp.storage is left untouched.

This also reworks how the storage selection is done when multiple
storages are defined but none is explicitly set, and what nuke limit we
use. Previously we will first try with a nuke limit of 0, and on failure
retry with the configured nuke limit. Furthermore, if we did not specify
the storage we will go through every one in RR fashion with a nuke limit
of 0 before retrying.

With this change we initialise beresp.storage with a storage backend from
the list before we enter v_b_r{}. At the end of v_b_r{} we will use
whatever storage is in beresp.storage. If the storage is NULL we fail
the request. In all cases we use the nuke limit.